[BUGFIX]  Fix fetching of the FlexForm settings in the list view plugin
The configuration array $this->conf has been initialized
with the settings of the list view plugin (e. g. showFilePreview).
If the setting of the search box plugin was empty (e. g.
if the checkbox showFilePreview was unchecked), the value
was not overwritten. Therefore, those settings did not have
any effect. This patch first fetches the FlexForm configuration
for the current plugin, then decides whether to overwrite it
with the FlexForm configuration from the search box plugin
and finally moves the configuration to $this->conf.