[FEATURE]  Add a column hidden_content and event ModifyFieldValuesBeforeStoringEvent
Add column "hidden_content" in the index table which is used for searching but not displayed in the result list.
This will be used for the new in-word-search feature of ke_search_premium but can also be used for other use cases (e.g. keywords, synonyms, word stemming). The new event ModifyFieldValuesBeforeStoringEvent can be used to add content to that column.