[FEATURE] Additional word characters
This patch adds a feature which allows to configure
additional word characters. These characters are
replaced during the search and in words containing
that characters during indexing. These modified
words are added as "hidden content" in the index.
That allows the user to search for words with
characters that would normally separate a word (like
dot or hyphen) like it would be a normal
character. This can be useful for example when
you search for serial numbers which contains a hyphen,
like in "AB-123".