[WIP][TASK] Remove dependency to AbstractPlugin, see #204
ke_search still relies on some of the functions
provided by the old "PiBase" class "AbstractPlugin"
which is planned to be removed in TYPO3 13.
This patch moves the relevant functions to a
"AbstractPlugin" class inside ke_search and removes
the dependency.
The usage of "pi_setPiVarDefaults" is removed.
Also, the class "Pluginbase" is refactored, is renamed
to "PluginBase" and now placed in the "Plugins"
namespace because that better fits the structure
than the "Lib" namespace.