[TASK] Improve logging for Indexer database actions, closes #59
Additionally, fix the "debugOnly"
parameter for inserting or updating.
This had no effect other
than outputting the result of the
query. Also, a different database
method had been used than the one
which would actually insert or
update the record. This has now been
fixed. If the debugOnly parameter is
set, now inserting or updating is
skipped and only the values are logged
(if LogLevel is set to "Debug").

Additionally, a part of the
errors is directly shown in
the indexing report.