[TASK] Migrate database queries
Includes removing of the getTableFields() method
in FilterOptionRepository.php. The method getSchemaManager()
is not available anymore and getTableFields() was
only used to check if the database columns do actually
exist in the filter option table. That's not necessary,
at least in ke_search itself. Third-party extensions
which use the create() function should check the fields
by themselves.