[TASK] Migrate docs to PHP-based rendering
The Documentation Team introduced the PHP-based rendering (guides) as replacement for the
previously used Sphinx-based rendering (which is now deprecated).

This change migrates the configuration from Settings.cfg to guides.xml.

To generate the docs, a Composer script is introduced:

    composer docs:generate

Additionally, a new GitHub Workflow has been added which checks, if there are warnings or errors
on rendering of the documentation.

See: https://typo3.org/article/transition-to-full-php-based-rendering-for-all-typo3-extension-manuals